cache implementation part - 2 using redis

+++ src/main/java/com/aishwarya/FinBank/dto/rules/RulesRequestDto.java
@@ -13,7 +13,7 @@ public class RulesRequestDto {
     @NotBlank(message = "Rule text cannot be blank")
     private String dslRule;
     @NotNull
-    private Integer loanTypeId;
+    private Long loanTypeId;
 
     public String getRuleText() {
         return dslRule;
